changed stop_container.sh

+++ scripts/stop_container.sh
@@ -1,4 +1,10 @@
 #!/bin/sh
 
+
 ContainerID = `docker ps | awk -F"" {print$1}` 
-sudo docker rm -f $ContainerID
+if [ -z "$ContianerID" ]; then
+    echo "No containers runnig"
+else
+    echo "A running container found removeing it : $ContainerID"
+    sudo docker rm -f $ContainerID
+fi
